+++ HISTORY.rst
@@ -10,6 +10,8 @@ Unreleased Changes
 * Changing default raster creation compression algorithm from LZW to DEFLATE,
   this is to address issues where we were seeing recreatable, but
   unexplainable LZWDecode errors in large raster data.
+* Fixed an issue that could cause the distance transform to be incorrect when
+  the sampling distance was in the noninclusive range of (0.0, 1.0).
 
 1.5.0 12/12/2018
 ----------------